The edit/remove buttons should not be submittable
Hello,
When the user has picked an image, the buttons to modify or remove this image appear. If, further down in the form, the user presses "Enter" to submit the form, the event is picked by the first submit button which is the vue-picture-input "modify image" button.
This PR fixes this behaviour by setting these buttons on the type "button" instead of the default "submit".
Thanks!